# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Agent": {
        "additionalProperties": false,
        "properties": {
          "connected": {
            "description": "A session is established. The environment is reported online once the session's jobs were reconciled.",
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "engineId": {
            "description": "Docker Engine ID.",
            "type": "string"
          },
          "environmentId": {
            "type": "string"
          },
          "hostname": {
            "description": "Engine host name reported at enrollment.",
            "type": "string"
          },
          "id": {
            "examples": [
              "0190a6e0-2222-7000-8000-000000000002"
            ],
            "type": "string"
          },
          "installId": {
            "description": "Generated by the agent once per state volume.",
            "type": "string"
          },
          "label": {
            "description": "Optional operator note.",
            "type": "string"
          },
          "lastConnectedAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastSeenAt": {
            "format": "date-time",
            "type": "string"
          },
          "revision": {
            "description": "Edit revision (the ETag).",
            "format": "int64",
            "type": "integer"
          },
          "revokedAt": {
            "format": "date-time",
            "type": "string"
          },
          "revokedReason": {
            "examples": [
              "removed"
            ],
            "type": "string"
          },
          "rotationPending": {
            "description": "A new credential waits for the agent's confirmation; the old one stays valid until then.",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "active",
              "revoked"
            ],
            "type": "string"
          },
          "transport": {
            "$ref": "#/components/schemas/AgentTransport"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "examples": [
              "0.0.0-edge"
            ],
            "type": "string"
          },
          "versionStatus": {
            "description": "outdated: previous minor release, still supported; upgrade it.",
            "enum": [
              "current",
              "outdated"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "environmentId",
          "label",
          "status",
          "connected",
          "installId",
          "engineId",
          "version",
          "versionStatus",
          "rotationPending",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "AgentTransport": {
        "additionalProperties": false,
        "properties": {
          "customCa": {
            "type": "boolean"
          },
          "managerUrl": {
            "type": "string"
          },
          "plainHttp": {
            "description": "The agent uses a plain-HTTP internal URL (DOCKYARD_MANAGER_ALLOW_HTTP); flagged on the host page.",
            "type": "boolean"
          }
        },
        "required": [
          "managerUrl",
          "plainHttp",
          "customCa"
        ],
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Stable snake_case error code; switch on this, not on message.",
            "examples": [
              "not_found"
            ],
            "type": "string"
          },
          "details": {
            "description": "Per-field problems, empty when not applicable.",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array"
          },
          "message": {
            "description": "Human-readable summary. Not stable; do not parse.",
            "examples": [
              "stack not found"
            ],
            "type": "string"
          },
          "requestId": {
            "description": "Request ID echoed in the X-Request-ID header and the manager logs.",
            "examples": [
              "4f1c2e7a9b0d4c3e8f6a1b2c3d4e5f60"
            ],
            "type": "string"
          },
          "retryable": {
            "description": "True when repeating the identical request later may succeed.",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "requestId",
          "retryable"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "description": "Input location, e.g. body.name, query.limit, path.stackId, header.If-Match.",
            "examples": [
              "body.name"
            ],
            "type": "string"
          },
          "message": {
            "description": "What is wrong with the value.",
            "examples": [
              "expected length >= 1"
            ],
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
...
      "PageAgent": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "description": "Items on this page (possibly empty, also when nextCursor is present).",
            "items": {
              "$ref": "#/components/schemas/Agent"
            },
            "type": "array"
          },
          "nextCursor": {
            "description": "Opaque cursor for the next page; absent on the last page.",
            "type": "string"
          },
          "total": {
            "description": "Number of items matching the filters that the caller may see, across all pages. Only on routes that document it.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "UpdateAgentInputBody": {
        "additionalProperties": false,
        "properties": {
          "label": {
            "description": "Operator note; empty clears it.",
            "maxLength": 200,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "bearerFormat": "DockYard API token",
        "description": "Scoped, expiring API token (#31) sent as Authorization: Bearer <token>. A token carries a subset of its owner's capabilities; each request is evaluated as token scope intersected with the owner's current effective permissions.",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieSession": {
        "description": "Browser session (#16). Set by POST /api/v1/auth/session as an HttpOnly, Secure, SameSite=Strict cookie. Unsafe methods authenticated by this cookie must come from the manager's own origin (Origin/Sec-Fetch-Site checks).",
        "in": "cookie",
        "name": "__Host-dockyard_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Public control API of the DockYard manager. All routes live under /api/v1 on the single public origin. Errors use the Error schema with media type application/problem+json. Every operation declares the capability it requires (x-dockyard-capability) and the scope at which it is checked (x-dockyard-scope). Browsers authenticate with the session cookie, other clients with a bearer API token. Conventions, errors, streams and versioning: docs/api/README.md in the DockYard repository.",
    "title": "DockYard API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/agents": {
      "get": {
        "description": "Agents newest first (active and revoked), filtered per item by agent.read.",
        "operationId": "list-agents",
        "parameters": [
          {
            "description": "Opaque cursor from a previous page's nextCursor. Only valid with the same filters and sort.",
            "explode": false,
            "in": "query",
            "name": "cursor",
            "schema": {
              "description": "Opaque cursor from a previous page's nextCursor. Only valid with the same filters and sort.",
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "description": "Maximum number of items to return.",
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "description": "Maximum number of items to return.",
              "format": "int64",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Only agents in these states.",
            "explode": false,
            "in": "query",
            "name": "status",
            "schema": {
              "description": "Only agents in these states.",
              "items": {
                "enum": [
                  "active",
                  "revoked"
                ],
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "description": "Only agents of this environment.",
            "explode": false,
            "in": "query",
            "name": "environmentId",
            "schema": {
              "description": "Only agents of this environment.",
              "maxLength": 64,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageAgent"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerToken": []
          }
        ],
        "summary": "List agents",
        "tags": [
          "Agents"
        ],
        "x-dockyard-capability": "agent.read",
        "x-dockyard-scope": "resource"
      }
    },
    "/api/v1/agents/{agentId}": {
      "delete": {
        "description": "Revokes the agent's credential at once and closes its live session (close code 4403); the environment stays offline and detached until an agent is enrolled with intent reattach:<environmentId>. Nothing on the host is touched. Requires If-Match. Removing an already revoked agent changes nothing.",
        "operationId": "delete-agent",
        "parameters": [
          {
            "description": "Agent ID.",
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "description": "Agent ID.",
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "description": "ETag of the revision being edited (from the resource's ETag header). Required: edits without it fail with 428 precondition_required; a stale value fails with 412 precondition_failed and the current ETag.",
            "in": "header",
            "name": "If-Match",
            "schema": {
              "description": "ETag of the revision being edited (from the resource's ETag header). Required: edits without it fail with 428 precondition_required; a stale value fails with 412 precondition_failed and the current ETag.",
              "maxLength": 1024,
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "412": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Precondition Required"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerToken": []
          }
        ],
        "summary": "Remove an agent",
        "tags": [
          "Agents"
        ],
        "x-dockyard-audit": "agent.remove",
        "x-dockyard-capability": "agent.remove",
        "x-dockyard-scope": "resource"
      },
      "get": {
        "operationId": "get-agent",
        "parameters": [
          {
            "description": "Agent ID.",
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "description": "Agent ID.",
              "maxLength": 64,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            },
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {
                  "description": "Strong entity tag of the returned revision; send it as If-Match when editing.",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerToken": []
          }
        ],
        "summary": "Get an agent",
        "tags": [
          "Agents"
        ],
        "x-dockyard-capability": "agent.read",
        "x-dockyard-scope": "resource"
      },
      "patch": {
        "description": "Sets the operator label. Requires If-Match.",
        "operationId": "update-agent",
        "parameters": [
          {
            "description": "Agent ID.",
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "description": "Agent ID.",
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "description": "ETag of the revision being edited (from the resource's ETag header). Required: edits without it fail with 428 precondition_required; a stale value fails with 412 precondition_failed and the current ETag.",
            "in": "header",
            "name": "If-Match",
            "schema": {
              "description": "ETag of the revision being edited (from the resource's ETag header). Required: edits without it fail with 428 precondition_required; a stale value fails with 412 precondition_failed and the current ETag.",
              "maxLength": 1024,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            },
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {
                  "description": "Strong entity tag of the returned revision; send it as If-Match when editing.",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "412": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Precondition Required"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerToken": []
          }
        ],
        "summary": "Update an agent",
        "tags": [
          "Agents"
        ],
        "x-dockyard-audit": "agent.manage",
        "x-dockyard-capability": "agent.manage",
        "x-dockyard-scope": "resource"
      }
    },
...
    "/api/v1/environments/{environmentId}/agents": {
      "get": {
        "description": "The environment's current and former (revoked) agents, newest first.",
        "operationId": "list-environment-agents",
        "parameters": [
          {
            "description": "Environment ID.",
            "in": "path",
            "name": "environmentId",
            "required": true,
            "schema": {
              "description": "Environment ID.",
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "description": "Opaque cursor from a previous page's nextCursor. Only valid with the same filters and sort.",
            "explode": false,
            "in": "query",
            "name": "cursor",
            "schema": {
              "description": "Opaque cursor from a previous page's nextCursor. Only valid with the same filters and sort.",
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "description": "Maximum number of items to return.",
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "description": "Maximum number of items to return.",
              "format": "int64",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageAgent"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerToken": []
          }
        ],
        "summary": "List an environment's agents",
        "tags": [
          "Environments"
        ],
        "x-dockyard-capability": "agent.read",
        "x-dockyard-scope": "environment"
      }
    },
...
  }
}